Properly end subroutine USB reader and close things
When we click on the "quit" button, the app will now close the USB
reader first, then close the port, then flush the influxdb writer, close
its socket, and quit the whole thing.

This allows to avoid weird issues. There will probably be a need for
another chan in order to close the DB writer loop, but so far it seems
to be OK thanks to the other closures.