Switch theme on the fly
This patch allows to switch the application them from "dark" to "light"
(and vice-versa) on the fly. It closes #9.

It also introduces some refactoring in the apolloUtils pkg, and adds the
capacity to dump (write) the configuration to a file.

The file can be either:
- the one passed with "-c" option
- ~/.gopolloplus

For the records, ~/.gopolloplus is the default location; in case you
don't have a configuration file, it will create it.